feat: Profiler Graph splitted in a cpp file, CMakeLists added as consequence
